requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
celery==5.4.0
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
requests==2.32.3
python-dotenv==1.0.1
joblib==1.4.2
astroquery==0.4.7
httpx==0.27.2
